+++ lib/annotate_rb/model_annotator/index_annotation.rb
@@ -5,6 +5,7 @@ module AnnotateRb
     module IndexAnnotation
       autoload :AnnotationBuilder, "annotate_rb/model_annotator/index_annotation/annotation_builder"
       autoload :IndexComponent, "annotate_rb/model_annotator/index_annotation/index_component"
+      autoload :Annotation, "annotate_rb/model_annotator/index_annotation/annotation"
     end
   end
 end
