+++ lib/annotate_rb/model_annotator/index_annotation/annotation_builder.rb
@@ -4,32 +4,6 @@ module AnnotateRb
   module ModelAnnotator
     module IndexAnnotation
       class AnnotationBuilder
-        class Annotation
-          HEADER_TEXT = "Indexes"
-
-          def initialize(indexes)
-            @indexes = indexes
-          end
-
-          def body
-            [
-              Components::BlankLine.new,
-              Components::Header.new(HEADER_TEXT),
-              Components::BlankLine.new,
-              *@indexes,
-              Components::LineBreak.new
-            ]
-          end
-
-          def to_markdown
-            body.map(&:to_markdown).join("\n")
-          end
-
-          def to_default
-            body.map(&:to_default).join("\n")
-          end
-        end
-
         def initialize(model, options)
           @model = model
           @options = options
